+++ src/kete/conversion.py
@@ -9,6 +9,27 @@ import numpy as np
 from numpy.typing import NDArray
 from . import _core
 from . import constants
+from ._core import compute_obliquity
+
+__all__ = [
+    "bin_data",
+    "compute_airmass",
+    "compute_albedo",
+    "compute_aphelion",
+    "compute_diameter",
+    "compute_earth_radius",
+    "compute_eccentric_anomaly",
+    "compute_h_mag",
+    "compute_obliquity",
+    "compute_semi_major",
+    "compute_tisserand",
+    "dec_degrees_to_dms",
+    "dec_dms_to_degrees",
+    "flux_to_mag",
+    "mag_to_flux",
+    "ra_degrees_to_hms",
+    "ra_hms_to_degrees",
+]
 
 
 logger = logging.getLogger(__name__)