+++ src/kete/rust/frame.rs
@@ -75,3 +75,37 @@ pub fn frame_change_py(
 pub fn wgs_lat_lon_to_ecef(lat: f64, lon: f64, h: f64) -> (f64, f64, f64) {
     geodetic_lat_lon_to_ecef(lat.to_radians(), lon.to_radians(), h)
 }
+
+/// Compute WCS84 Geodetic latitude/longitude/height from a ECEF position.
+///
+/// This returns the lat, lon, and height from the WGS84 oblate Earth.
+///
+/// Parameters
+/// ----------
+/// x :
+///     ECEF x position in km.
+/// y :
+///     ECEF y position in km.
+/// z :
+///     ECEF z position in km.
+#[pyfunction]
+#[pyo3(name = "ecef_to_wgs_lat_lon")]
+pub fn ecef_to_wgs_lat_lon(x: f64, y: f64, z: f64) -> (f64, f64, f64) {
+    ecef_to_geodetic_lat_lon(x, y, z)
+}
+
+/// Calculate the obliquity angle of the Earth at the specified time.
+///
+/// This is only valid for several centuries near J2000.
+///
+/// The equation is from the 2010 Astronomical Almanac.
+///
+/// Parameters
+/// ----------
+/// time:
+///     Calculate the obliquity angle of the Earth at the specified time.
+#[pyfunction]
+#[pyo3(name = "compute_obliquity")]
+pub fn calc_obliquity_py(time: f64) -> f64 {
+    calc_obliquity(time).to_degrees()
+}