+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Epoch and Perihelion time conversion when loading JPL Horizons covariance matrices was
+  not being done for UTC to TDB, leading to small residuals in covariance sampling.
+
+
 ## [v1.0.0]
 
 ### Added